Add methods to locate and vet state references. Introduced caching for prototyped methods.
Prototyped .indexOf() and .vetIndexOf(), for retrieving the index of a valid state reference, or -1 to indicate otherwise.
Supporting functions handle the recursion and iteration... Trying genData as parser for tokenized state reference strings. Codebase is larger - may revert in near future.
As with the stable version, these string-parsers use caching. The caching identifies when the string/query is global or specific to the current state.